+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ## NEXT ()
 
+* [new] Added column selector to members listing
 * The `mb:user:change_password_form` tag has been deprecated
 * Updated templates and documentation to use new Statamic tags
 * Updated templates to use runtime parser syntax
-* Include file field example in profile template
+* Include file field example in the profile template
 
 ## 1.0.4 (2022-11-28)